refactor: move query types to query_handler
Moves types that are only used for handling queries to the query_handler
module.
